requirements.txt
streamlit>=1.32.0
primer3-py>=2.0.0
biopython>=1.83
dna-features-viewer>=3.1.2
matplotlib>=3.8.0
openpyxl>=3.1.2
pandas>=2.2.0
reportlab>=4.1.0
Pillow>=10.2.0
bcrypt>=4.0.0
requests>=2.31.0
